+++ common/src/main/java/ru/ancndz/bandagez/config/ModConfiguration.java
@@ -2,7 +2,6 @@ package ru.ancndz.bandagez.config;
 
 import java.util.HashMap;
 import java.util.Map;
-import java.util.function.Function;
 
 public abstract sealed class ModConfiguration<T> permits ServerModConfiguration, ClientModConfiguration {
     private static ClientModConfiguration<?> client = null;
@@ -10,21 +9,17 @@ public abstract sealed class ModConfiguration<T> permits ServerModConfiguration,
 
     private final Map<String, T> values = new HashMap<>();
 
-    private <TT extends T> TT getValueRaw(String path) {
+    protected <TT extends T> TT getValueRaw(String path) {
         return (TT) this.values.get(path);
     }
 
-    protected <V extends Comparable<? super V>> V getValue(String path) {
-        return this.<V>getConverter().apply(getValueRaw(path));
+    protected Map<String, T> getValues() {
+        return this.values;
     }
 
-    protected <V extends Comparable<? super V>> void putValue(ConfigEntry<V> configEntry) {
-        values.put(configEntry.getPath(), this.<V>getValueConverter().apply(configEntry));
-    }
+    protected abstract <V extends Comparable<? super V>> V getValue(String path);
 
-    protected abstract <V extends Comparable<? super V>> Function<? extends T, V> getConverter();
-
-    protected abstract <V extends Comparable<? super V>> Function<ConfigEntry<V>, ? extends T> getValueConverter();
+    protected abstract <V extends Comparable<? super V>> void putValue(ConfigEntry<V> configEntry);
 
     public static void setClientConfig(ClientModConfiguration<?> clientModConfigurationConfig) {
         ModConfiguration.client = clientModConfigurationConfig;
